Took out the toolbar for the Login Page. Deleted Musician Profiule.axml. Put Toolbar on Profile and Search Page. Styled Search page, spinners and edit text still needs to be styled.
