+++ Gigue_Xamarin_Android/Gigue/Resources/layout/ListItemRow.axml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
+    android:background="#242424"
     android:minHeight="50dp">
     <TextView
         android:id="@+id/txtFirstName"
@@ -11,7 +11,7 @@
         android:lineSpacingExtra="3dp"
         android:paddingLeft="10dp"
         android:paddingTop="5dp"
-        android:textColor="#fff"
+        android:textColor="#ffffff"
         android:textStyle="bold"
         android:typeface="sans" />
     <TextView
@@ -22,7 +22,7 @@
         android:lineSpacingExtra="3dp"
         android:paddingLeft="10dp"
         android:paddingTop="5dp"
-        android:textColor="#fff"
+        android:textColor="#ffffff"
         android:textStyle="bold"
         android:typeface="sans" />
     <TextView
@@ -30,21 +30,19 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_below="@id/txtFirstName"
-     
         android:paddingLeft="10dp"
         android:paddingTop="5dp"
-        android:textColor="#fff"
+        android:textColor="#ffffff"
         android:textStyle="bold"
         android:typeface="sans" />
-  <TextView
+    <TextView
         android:id="@+id/txtCity"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_below="@id/txtLastName"
-       
         android:paddingLeft="10dp"
         android:paddingTop="5dp"
-        android:textColor="#fff"
+        android:textColor="#ffffff"
         android:textStyle="bold"
         android:typeface="sans" />
 </RelativeLayout>
