--- Gigue_Xamarin_Android/Gigue/Resources/layout/musicianProfile.axml
@@ -1,44 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:orientation="vertical"
-    android:minWidth="25px"
-    android:minHeight="25px"
-    android:layout_width="fill_parent"
-    android:layout_height="fill_parent"
-    android:id="@+id/linearLayout1"
-    android:background="#424242">
-    <ScrollView
-        android:minWidth="25px"
-        android:minHeight="25px"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:id="@+id/scrollView1">
-        <TableLayout
-            android:layout_width="fill_parent"
-            android:layout_height="fill_parent"
-            android:stretchColumns="1">
-            <TextView
-                android:text="James Kulon"
-                android:layout_width="match_parent"
-                android:layout_height="43.0dp"
-                android:id="@+id/textView1"
-                android:gravity="center"
-                android:textSize="30dp"
-                android:textStyle="bold"
-                android:typeface="serif"
-                android:textColor="#f2854b" />
-            <ImageView
-                android:src="@drawable/james"
-                android:layout_width="match_parent"
-                android:layout_height="254.0dp"
-                android:id="@+id/imageView1" />
-            <Switch xmlns:app="http://schemas.android.com/apk/res-auto"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:id="@+id/switch1"
-                android:textOn="Available for Work"
-                android:textOff="Not Available"
-                app:trackTint="#00000000" />
-        </TableLayout>
-    </ScrollView>
-</LinearLayout>
